refactor(rag): route remaining RagStatus imports through app.lib.rag
Completes the public-boundary enforcement: rag_mcp and app.models.drive now
import RagStatus from app.lib.rag instead of app.rag.enums.

app.models.drive could not import the facade directly (RagStatus is a field
annotation needed at class-definition time, and app.lib.rag -> app.rag.store
-> app.models.drive formed a cycle). Resolved by deferring store.py's only
back-edge — the DriveFile import used inside store_and_link_chunks — to a
lazy import, with an explanatory note. The facade's top-level imports stay
intact, preserving the tests/lib patch targets.
