+++ app/rag_mcp/tools.py
@@ -6,8 +6,8 @@ from sqlmodel import col, func, select
 
 from app.lib.db import session_scope
 from app.lib.log import get_logger
+from app.lib.rag import RagStatus
 from app.models.drive import DriveFile
-from app.rag.enums import RagStatus
 from app.rag.models import DocumentChunk
 from app.rag.utils import resolve_source_name
 from app.rag_mcp.schemas import ChunkStats, DocumentSection, FileInfo, FileMetadata, SectionKey
